Reverted database format changes for Compact Sqlite and creation of a BaseWikiData class.
The reason for the different table names "wikiwords" and "wikiwordcontent" is
that the former only contains meta information of a wiki page in
Original Sqlite database while the latter also contains the page content
of Compact Sqlite database.

A future idea is to replace the "wikiwordcontent" table by a
"wikiwords" table for the metadata only and to place the content in the
"datablocks" table for Compact or refer to "datablocksexternal" for Original.

The goal is to use a common data format and
WikiData class for both storage variants with only a different default setting if
the wikipage datablocks should be stored internally or as files.